_requirements.txt
altgraph==0.17
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
cimport==0.3.1rc0
Cython==0.29.34
et-xmlfile==1.1.0
future==0.18.2
idna==3.4
importlib-metadata==4.6.3
jieba==0.42.1
jsonschema==4.17.3
lxml==4.6.3
openpyxl==3.0.7
pefile==2021.5.24
Pillow==8.3.1
pyDes==2.0.1
pyinstaller==4.5
pyinstaller-hooks-contrib==2021.2
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.19.2
python-docx==0.8.11
pywin32==305
pywin32-ctypes==0.2.0
requests==2.30.0
typing-extensions==3.10.0.0
urllib3==2.0.2
WMI==1.5.1
xlrd==2.0.1
xlwt==1.3.0
xpinyin==0.7.6
zipp==3.5.0
